refactor: Dao layer eliminated, packages restructured
