+++ src/main/java/com/codecool/freshcucumbersbackend/FreshCucumbersBackendApplication.java
@@ -1,6 +1,6 @@
 package com.codecool.freshcucumbersbackend;
 
-import com.codecool.freshcucumbersbackend.service.TopTenMovieLoader;
+import com.codecool.freshcucumbersbackend.service.TopTenMovieHandler;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -11,7 +11,7 @@ import javax.annotation.PostConstruct;
 public class FreshCucumbersBackendApplication {
 
     @Autowired
-    private TopTenMovieLoader topTenMovieLoader;
+    private TopTenMovieHandler topTenMovieHandler;
 
     public static void main(String[] args) {
         SpringApplication.run(FreshCucumbersBackendApplication.class, args);
@@ -19,7 +19,7 @@ public class FreshCucumbersBackendApplication {
 
     @PostConstruct
     public void afterInit() {
-        topTenMovieLoader.populate();
+        topTenMovieHandler.populate();
 
     }
