+++ src/main/java/com/codecool/freshcucumbersbackend/config/AppConfig.java
@@ -1,13 +1,13 @@
 package com.codecool.freshcucumbersbackend.config;
-import com.codecool.freshcucumbersbackend.dao.TMDbApiHandler;
+import com.codecool.freshcucumbersbackend.service.TMDbApiHandler;
 import com.codecool.freshcucumbersbackend.entity.Review;
-import com.codecool.freshcucumbersbackend.service.MovieCreator;
+import com.codecool.freshcucumbersbackend.repository.MovieCreator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
 
 @Configuration
-public class CreatorConfig {
+public class AppConfig {
 
     @Bean
     public MovieCreator creator() {
